Allow use of pylint installed in virtualenv
The `inithook` method returns immediately if it detects that `pylint` is
running inside a virtualenv. This is the typical use case.

However, there are cases when the `pylint` itself is installed in a
virtualenv but is being called from a different virtualenv. For example,
I am using [`coc-python`](https://github.com/neoclide/coc-python) and I
have a separate `pyenv-virtualenv` for `neovim` residing in
`~/.pyenv/versions/neovim`. I have `pip3 install jedi pylint black` in
that virtualenv and I have hardcoded those paths in `coc-settings.json`.
In this case, it is undesirable for `inithook` to return immediately.

This commit adds `ignore_pylint_venv=False` parameter to `inithook`
method. When the argument is `True` the method does _not_ return
immediately but proceeds with rest of the `inithook` logic.

This commit updates the documentation showing how to pass the argument
from a command line.