Make the automated test work without which(1)
which(1) is a third-party executable that's not installed on every
system out there, and it is entirely redundant to shell builtins such
as `command -v` (POSIX sh) or `type -P` (bash).  Since there does not
seem to be any real need to use full path here, just inline the `pylint`
call.